requirement.txt
Django==1.11
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.4
rest_framework_custom_viewsets==0.0.4
